Add an API to load default configuration values (#13788)
Currently we have RedisModule_LoadConfigs which the module is expected
to call during OnLoad which sets the configuration values from the
config queue or it sets the default value.

The problem is that the module might still want to support loading
values from the command line. If we want to give precedence to the
config file values then it means the module needs to set the values
before calling the Load Config function.

The problem is that then the API overrides the variables which were set
from the module command line with default values.

The new API should solve that in the following way.

1.Module registers its configuration parameters with redis 2.Module
calls RedisModule_LoadDefaultConfigs which loads the default values for
all the registered configuration parameters of the module 3.Module sets
the variables internally using the values it got from the command line
4.Module calls RedisModule_LoadConfigs which will set the values based
on the redis configuration file.

This allows for the default values to be set, for the module to override
them and for redis to override what the module wrote. In short it
determines a logical flow and ordering of where the values for the
parameters should come from.

The change done by all these previous commits:
a91d69228
df8de8507
9faed2fe5
0d2e5c3e2
7836d856a
3a94c45c3
c6adcac35
e57564825
350036a5e
67212f2b8
4dff19eae
350c4c054
b95b833aa
196166f49
708e8ded5
5ec1def1c
5c31cb919
840786f77
e6671c222

Co-authored-by: YaacovHazan <yaacov.hazan@redislabs.com>